# Optimize pin reconfiguration in DSHOT ISR for H7