--- a/src/main/drivers/io.h
+++ b/src/main/drivers/io.h
@@ -131,3 +131,6 @@ void IOInitGlobal(void);
 typedef void (*IOTraverseFuncPtr_t)(IO_t io);
 
 void IOTraversePins(IOTraverseFuncPtr_t func);
+
+GPIO_TypeDef* IO_GPIO(IO_t io);
+uint16_t IO_Pin(IO_t io);