I found that I would sometimes get DSHOT telemetry errors with an H750 based flight controller and octo copter configurations. Upon investigation, I found that the headroom between the last call of `motor_DMA_IRQHandler` and the DSHOT telemetry response was only about 4 us. 

This shows the timing. The debug pin is H during the execution of the `motor_DMA_IRQHandler`
![image](https://user-images.githubusercontent.com/647005/101062576-a8265780-3546-11eb-812f-3ba2c4997922.png)

In order to fix this problem, I optimized the pin-reconfiguration in the `pwmDshotSetDirectionInput` function. I first thought that the re-configuration may not even be needed, but @hydra pointed out the glitch that happens during the timer configuration (see #8836). To reduce the time spent in the ISR, I use LL library calls instead. 

With the new code, the ISR takes about 1.6 us (from 2.7 us) and the headroom increases from 4 us to about 12 us, which avoids the DSHOT telemetry errors.

![image](https://user-images.githubusercontent.com/647005/101063308-837eaf80-3547-11eb-8da1-10e04f66e845.png)

Also, I confirmed that the glitch at the output still doesn't happen. 

![image](https://user-images.githubusercontent.com/647005/101063572-cb053b80-3547-11eb-8d09-e7061e8cf465.png)