@XING-IF 2020-12-24
I made a "discovery" with the iFlight BEAST H7 AIO flight board as well. First of all, the tested JazzMaverick version 16.71 and 16.9 worked until BF 4.2.5, no dshot error, RPM filter working, RPM reading and motor spinning (No in-flight issues). Tried to flash latest JESC firmware, motor3 was suddenly at 100% dshot error, no rpm reading anymore. After further investigation and a hint of Steve Evans did we analyze the board and tried to fix an error.

Nonetheless, I have tested several screnarios:
1.) 4.2.5 and JazzMaverick (Blheli-M) worked, no visible issues whatsoever.
2.) 4.2.5 and JESC didn't work, motor3 100% Error, no RPM, not spinning.
3.) 4.2.6 I think release Nr.35, Blheli-M and JESC both delivered same errors.
4.) Deactivating dshot_bitbang let the motors spin, but no rpm reading.
5.) BF 4.3.0 (20-12-23) both ESC firmwares, same errors.

The new board version will be in my hands soon, I'm able to provide one for testing as well. (iFlight BEAST H7).
The question is, what did change with 4.2.6 and above so motor3 has 100% error?
We are still not 100% sure it's a hardware issue. How can we fix this issue on the software side, falshing 4.3.0 but using the 4.2.5 working dshot settings? BF 4.3.0 currently development versions are also necessary to use the new Barometer!
Thanks a lot guys.