@mikeller 2020-12-24
@mluessi: For context, the only difference pertaining to H7 and / or Dshot between the working / non-working versions in the maintenance branch is this pull request. I am currently contemplating removing it from the maintenance branch for the next release, to buy us some time to get to the bottom of this.