@kolaa 2021-01-05
Same issue on iFlight BEAST H7 AIO with JESC firware
biredectional DSHOT didn't works on motor3 on 4.2.4, 4.2.5 firmwares