@hydra 2021-02-02
@XING-IF @kolaa @mikeller If you have the H7 boards from IFlight, suggest running a git bisect between 4.2.5 and 4.2.6 to confirm the offending commit is the one from this PR, otherwise it's just a guess and might not be the root cause.  Can someone confirm this has been done?