@mluessi 2021-02-05
The changes in this PR are fairly minimal and as far as I can tell the LL library calls should be doing the same as the as the previous code. So it is surprising that it is causing problems.

I echo what @hydra said. Someone who is experiencing this problem should bisect it and see if this is the actual offending commit. I have not been able to reproduce the problem myself.