@mikeller 2021-02-09
@SteveCEvans was investigating this, and I think he was able to pin this down to a problem with the iFlight hardware. So if he can confirm this we can mount another attempt at cherry-picking this into 4.2.