@SteveCEvans 2021-02-09
I can confirm that 4.2.5 with JESC reports 100% errors on motor 3. Jazz 16.71 appears to work with bidirectional DSHOT600.
Contrary to the above, however, I find that 4.2.6 behaves the same way.
With current master (2467b32) motor 3 does not report errors, nor does it spin. Unless...

Now to dig a little deeper into the hardware as hinted at by @XING-IF above.

Having hunted around for reference BLHeli designs using the [EFM8BB21F16G](https://www.silabs.com/documents/public/data-sheets/efm8bb2-datasheet.pdf) I found the package pinout:

![image](https://user-images.githubusercontent.com/11480839/107414197-972b3100-6b09-11eb-8e72-ff1e87ccf9c8.png)

And then corresponding to the G-H-10 ESC definition I got the G pinout which I found [HERE](https://github.com/bitdump/BLHeli/blob/master/BLHeli_S%20SiLabs/G.inc). If I've interpreted this correctly then the following excerpt defines the pin usage.

```
;**** **** **** **** ****
; Bootloader definitions
;**** **** **** **** ****
RTX_PORT		  		EQU 	P0		; Receive/Transmit port
RTX_MDOUT				EQU	P0MDOUT	; Set to 1 for PUSHPULL				
RTX_MDIN				EQU	P0MDIN	; Set to 1 for DIGITAL				
RTX_SKIP				EQU	P0SKIP	; Set to 1 for SKIP				
RTX_PIN					EQU 	5		; RTX pin
```

So, correct me if I'm wrong, but P0, pin 5, would appear to be the IO pin for serial comms. Below is an image of two of these on the AIO with series resistors connected to pin 0.5 highlighted.

![IMG_4929](https://user-images.githubusercontent.com/11480839/107411670-ba081600-6b06-11eb-9186-4bbcf38f7676.jpeg)

All the same one might expect. On three of the ESCs they are 1K, and on one (interestingly not motor 3!) it's 4K7.

In the absence of any schematics I thought it might be interesting to hook this up line for motor 3 to a logic analyser.

![IMG_5071](https://user-images.githubusercontent.com/11480839/107417113-2ab23100-6b0d-11eb-95b5-3b1e5ad54cbd.jpeg)

And, in best hardware debugging tradition, as soon as you make a connection to non-portable test equipment, things start to work. Both Jazz 16.71 and JESC work with the logic analyser connected.

Jazz 16.71
![image](https://user-images.githubusercontent.com/11480839/107418670-d740e280-6b0e-11eb-8ed1-fee4726eca12.png)
![image](https://user-images.githubusercontent.com/11480839/107419715-0146d480-6b10-11eb-9def-451aca452006.png)

JESC 2.3
![image](https://user-images.githubusercontent.com/11480839/107418773-f770a180-6b0e-11eb-9d80-2f18b8ec13a5.png)
![image](https://user-images.githubusercontent.com/11480839/107419782-16bbfe80-6b10-11eb-9713-c380c64e6015.png)

Interesting the JESC only returns telemetry data 2 or 3 out of every 4 updates (once the motors are running), whereas Jazz returns them every time. Also the turnaround time is ~18us on Jazz vs ~28us on JESC.

The fact that connecting the logic analyser fixes this rather demonstrates that there's a signal integrity issue here.