@hydra 2021-02-10
@SteveCEvans Great digging!

1) Did the iFlight guys give the PCB assembly people the wrong BOM/component placements or did they use a different resistor on one of the ESC MCU's intentionally, and if so why?

2) Where does the other end of the resistor go?  To the FC's motor output pin? Anywhere else?

3) Do you have a scope and probes that will allow you to get traces of the signals?  This would shed some light on things.

4) When you removed/disconnected your logic analyzer did the ESCs stop working again?

Given the fact that neither I nor @mluessi have been able to replicate the issue, and that it's only confirmed on iFlight hardware, and that they've also done new board revisions and that your investigation appears to confirm some signal integrity issue on the iFlight hardware I think the PR should be re-instated.

Certainly the code in this PR *has* fixed issues for SP Racing customers who were experiencing issues with recent BF firmware.  I've had no reports of ESC DSHOT telemetry issues from customers since this PR was added and since the BLHeli timer clock bug was fixed.