@SteveCEvans 2021-02-10
(1) certainly looks like a pick-n-place error, but I can’t confirm that.

(2) to a blind via. Thereafter, on a board using BGA packages your guess is as good as mine!

(3) yes, but given (1) is it really worth the effort?

(4) yes.

I’ve flown https://github.com/betaflight/betaflight/pull/10525 on an SPRACINGH7ZERO with T-Motor  F55A ESC without issue, but that’s 4.3 rather than 4.2. Not sure if this PR is needed TBH, but could be ESC dependant?