@hydra 2021-02-11
> (3) yes, but given (1) is it really worth the effort?

I would say so yes, so that we can characterize the failure.

Certainly does look like signal integrity though given (4) above.

With regards to setups being flown without this PR, yes, people have done so but not everyone.  Still there's no denying that the CPU cycles saved in this key area makes the DSHOT telemetry implementation more robust and gives performance benefits.

It could be ESC dependent, as it depends if the ESC firmware and hardware is to-spec.  All the ESC MCUs and the FC MCUs have different timings for transitioning a GPIO pin from output to input and back again, so having the FC ready to receive earlier allows the FC to cope with ESCs that can do it faster, which isn't a bad thing and fixes DSHOT for the Octocopter use-case @mluessi had issues with.